Fix calculated vote tally
If the input text field for a stance remained empty, a single vote for this position was faultily added nonetheless. The tally calculation has been fixed to explicitly check whether the given vote list for a stance is in fact empty, and to remain at 0 votes for that stance if so.